Frontend rules (#467)
Frontend rules specifies a set of rules that should be applied in
HAProxy frontend configuration. The set of keywords are from here
https://cbonte.github.io/haproxy-dconv/1.7/configuration.html#4.1
Only frontend sections can be applied here. It is up to user to
provide valid set of rules.
This allows acls or other options in frontend sections in HAProxy config.
Frontend rules will be mapped with Ingress Rules according to port.